+++ src/json_parser.cpp
@@ -3,8 +3,8 @@
   will be useful, but WITHOUT ANY WARRANTY; without even the implied warranty
   of MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. 
   
-  This is the sample code for Leopard USB3.0 camera, mainly for reading from 
-  config.json at camera startup to perform group register writes and captures.
+  This is the sample code for Leopard USB3.0 camera, mainly for loading from 
+  config.json to perform group register writes and captures.
 
   Author: Danyu L
   Last edit: 2019/06
@@ -44,7 +44,7 @@ void json_parser(int fd, char *json_buffer)
 
     char *dev_name;
     uint8_t sub_addr;
-    int addr_width, val_width;
+    uint8_t addr_width, val_width;
     uint32_t reg_addr = 0;
     uint32_t reg_val = 0;
     int execute_flg = 0;