+++ CHANGELOG.md
@@ -7,3 +7,8 @@ Initial release version
 
 ### v.1.1 - 2019-05-02
 Add tightly packed RAW8 support
+
+### v.1.2 - 2019-06-21
+Add firmware erase
+Add "Load from json" for register setting write, image capture functions
+